Updated the driver
Updated the driver and changed a lot. 
- Starting with switch(function) and the end of it. 
- Changed the json output code a lot to more visibility and a better way of working. 
- Changed the web sever outout so that we need less code.

In #ifdef USE_DOMOTICZ i have changed nothing because i don't have a domotica system.
Therefor someone with a stable running system can build it further or we can work together on it.

Dirver version changed to v1.0.0.2
Theo, i hope the if (11 == (uptime %100)) { .. part is ok. Please have a look on it.